fix: correctness bugs found by static audit
- FilePath::makeRelativeTo: reorder while-condition so iterator bounds
  are checked before dereference (UB on fully-shared path prefixes).
- SearchIndex::scoreText: compute the unmatched-letter gap in signed
  arithmetic; the previous size_t*(-1) multiply wrapped and corrupted
  fuzzy-search ranking for any non-adjacent match.
- BookmarkCache: initialize m_nodeBookmarksValid/m_edgeBookmarksValid;
  an indeterminate-true read could skip loading bookmarks from storage.
- Graph: initialize m_hasTrailOrigin (read by hasTrailOrigin()).
- UndoRedoController: initialize m_historyOffset in the constructor.
- TaskGroupParallel/TaskGroupSequence: initialize scheduling state
  members in the constructor rather than relying on doEnter() ordering.

+++ src/lib/data/search/SearchIndex.cpp
@@ -444,8 +444,9 @@ int SearchIndex::scoreText(const std::wstring& text, const std::vector<size_t>& 
 		// unmatched and consecutive
 		if (i > 0)
 		{
-			unmatchedLetterScore += static_cast<int>(
-				(indices[i] - indices[i - 1] - 1) * unmatchedLetterBonus);
+			unmatchedLetterScore +=
+				(static_cast<int>(indices[i]) - static_cast<int>(indices[i - 1]) - 1) *
+				unmatchedLetterBonus;
 			consecutiveLetterScore += (indices[i] - indices[i - 1] == 1) ? consecutiveLetterBonus : 0;
 		}
